requirements.txt
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
langchain-core>=0.1.0
langchain-gigachat>=0.1.0
urllib3>=1.26.0